Change socket types for communication between GUI and server
Any request (change slice, etc.) from the GUI is communicated through
a PAIR socket. It was PUB previously. Since there won't be more
than one server and we don't wont more than one GUI client running at
the same time, the PAIR socket is apparently a better choice.

The server will send data to the GUI via a REP socket. In principle,
a PUSH socket can also be used here. But in case of a network delay,
the GUI client will always receive the old update instead of the latest
one.

Last but not least, now the server is a ZMQ server and the GUI is a
ZMQ client.